# api/docs/telephant-api-documentation.yml
openapi: 3.0.0
info:
  title: Telephant API
  description: |
    *The mechanism for swapping airtime is powered by Telephant's APIs. Developers can use the API to tap into the platform's capabilities and incorporate the service into their own apps.*

    The Telephant API's backbone is HTTP. The nature of our API is RESTful. It has a set of predictable resource URLs. It returns HTTP response codes to indicate errors. It also accepts JSON as an HTTP body payload and returns it, along with some other useful formats.
    
    You can use your favorite HTTP/REST library for your programming language or an HTTP client like Postman or CURL to use Telephant's API.

    Happy programming **:-)**

    ## HTTP Response Status Codes
    HTTP response status codes show whether or not a particular HTTP request was completed successfully. Section 10 of RFC 2616 defines the status codes listed below. RFC 7231 contains an updated specification.

    Telephant API makes use of the following HTTP status codes, grouped into three classes:

    ### Successful Responses - 2XX

    #### 200 - OK

    The request has succeeded.

    #### 201 - Created

    As a result of the request's success, a new resource has been generated. This is the response that is normally sent in response to POST or some PUT queries.

    ### Client Error - 4XX
    
    #### 400 - Bad Request
    Due to incorrect syntax, the server was unable to understand the request.
    
    #### 401 - Unauthorized
...
    
...
    
...
    
...

...

...

...

...

...
servers:
...
paths:
  /v1/accounts:
    get:
      summary: Get account(s).
      description: Get the details of a specific account(s). List one or all of the details of accounts available to this username.
      tags:
        - Accounts
      security:
        - BearerAuth: []
      parameters:
      - in: header
        name: account_id
        description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
        schema:
          type: string
          format: uuid
        required: false
      - in: query
        name: page
        description: The page number of results to return.
        schema:
            type: integer
      - in: query
        name: limit
        description: The numbers of items to return per page.
        schema:
          type: integer
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              examples:
                Accounts-Get:
                  $ref: '#/components/examples/Accounts-Get'

        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'
    post:
      summary: Create account.
      description: Create a new group account.
      tags:
        - Accounts
      security:
        - BearerAuth: []
      requestBody:
        description: Define the properties of the account to create.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accountName
              properties:
                accountName:
                  type: string
                  minLength: 3
                  maxLength: 16
                  example: my-company
                  description: |
                                  This field must be used as a reference on bank statements, 
                                  and Telephant will use it to simplify the process.
                                  
                                  **Note:** Only alphanumeric characters and hyphens as spaces are accepted 
                                  in this field.

                                  **Note:** The account id in the header will be ignored because 
                                  the owner of the group account will be set to your username's phone number.

                  format: '^[a-z0-9-]{3,16}$'
                description:
                  type: string
                  maxLength: 300
                  example: My Company (Pty) Ltd Airtime Account

      responses:
        '200':
          description: Success
          content: 
            application/json:
              example:
                {
                    message: 'The group account was successfully created.'
                }
        '400':
          description: BadRequestError
          content: 
            application/json:
              example:
                {
                    "code": 400,
                    "name": "BadRequestError",
                    "message": "Account validation failed : A account name is required."
                }

        '409':
          description: Conflict
          content: 
            application/json:
              example:
                {
                    "code": 409,
                    "name": "ConflictError",
                    "message": "This group account name is already registered.  Please choose another."
                }

        '5XX':
          $ref: '#/components/responses/5XX'
    put:
      summary: Update account.
      description: |
                      Update the details of a group account.

                      **Note:** It is not possible to update your own username's account through the API 
                      so an account_id must be provided in the header.
      tags:
        - Accounts
      security:
        - BearerAuth: []
      requestBody:
        description: |
                        Define the properties of the account to update.

                        The fields that are available to be modified are:
                        - owners
                        - description
                        - businessType
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owners:
                  type: string
                  pattern: '^(27)[0-9]{9}$'
                  example: 27821234567
                  description: | 
                    In the South African phone number format, a user's unique mobile phone number.

                    You can assign multiple owners to a group account but only one per request.
                description:
                  type: string
                  maxLength: 300
                  example: My Company (Pty) Ltd Airtime Account
                businessType:
                  type: string
                  enum:
                    - private
                    - public
                    - ngo
                    - gov
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '201':
          description: Success
          content: 
            application/json:
              example:
                {
                    message: 'The group account was successfully updated.'
                }          

        '4XX':
          description: BadRequestError
          content:
            application/json:
              examples:
                ConflictError-1:
                  value:
                    {
                        "code": 400,
                        "name": "ConflictError",
                        "message": "The user telephone number is already an owner on this account."
                    }
                ConflictError-2:
                  value:
                    {
                        "code": 400,
                        "name": "ConflictError",
                        "message": "Not a valid user's telphone number."
                    }
                ConflictError-3:
                  value:
                    {
                        "code": 400,
                        "name": "ConflictError",
                        "message": "No relevant fields to update were defined."
                    }

        '5XX':
          $ref: '#/components/responses/5XX'
    delete:
      deprecated: true
      summary: Delete account.
      description: |
                      Delete a specific group account.

                      **Note:** The ability to delete a group account has been removed.  
                      In future a method for achieving the deletion of an account will be implemented
                      on the UPDATE method.
      tags:
        - Accounts
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '410':
          description: Gone
          content: 
            application/json:
              example:
                {
                    "code": 410,
                    "name": "Gone",
                    "message": "The ability to delete a group account has been removed.  In future a method for achieving this will be implemented."
                }

#### AIRTIME ####
...
  /v1/notifications/count:
    get:
      summary: Get a count of notifications.
      description: Get a count of read, unread and total notifications.
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              examples:
                Notifications-Get:
                  $ref: '#/components/examples/Notifications-Count'
        '4XX':
          $ref: '#/components/responses/4XX'
        '5XX':
          $ref: '#/components/responses/5XX'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    4XX:
      description: |
        Client error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 400
                description: |
                  Please review the documentation for [HTTP Response Status Codes](/#section/HTTP-Response-Status-Codes)

                  Possible values are:
                    - 400 - Bad Request
                    - 401 - Unauthorized
                    - 403 - Forbidden
                    - 404 - Not Found
                    - 405 - Method Not Allowed
              name:
                type: String
                example: 'BadRequest'
                description: This value will always be returned as 'bumpy case'.
              message:
                type: String
                description: |
                  The message field will provide specific information as to the reason for the error.
                example: 'The account ID provided is not valid.'
    5XX:
      description: |
        Server error  
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 500
              name:
                type: String
                example: 'InternalServerError'
              message:
                type: String
                example: 'We are sorry but something went wrong on our side.'
  examples:
...
    Notifications-Count:
      value:
        {
            "read": 10,
            "unread": 2,
            "total": 12
        }
...
    Accounts-Get:
      value:
        {
          "page": 1,
          "total_pages": 1,
          "total_results": 3,
          "results": [
                        {
                          "_id": "60f57f3a6023f819ae7aa454",
                          "accountName": "27829524031"
                        },
                        {
                          "_id": "60ebd53e2f79311541618bc4",
                          "accountName": "MyCompany",
                          "description": "The account description for My Company"
                        },
                        {
                          "_id": "60ebd53e2f79311541618bc5",
                          "accountName": "My Company 2",
                          "description": "The account description for My Company 2"
                        }
                    ]
                }
...